% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Report the current time}
\usage{
what_time(language = "fr")
}
\arguments{
\item{language}{the language that you prefer the output to be c("en", "fr")}
}
\value{
A string indicating the time
}
\description{
Report the current time
}
\section{Functions}{
\itemize{
\item \code{what_time()}: we put more details about the funciton in this pacakge

}}
\examples{
what_time()
what_time(language = "en")
}
